% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{board_colors}
\alias{board_colors}
\title{Construct a board color preset}
\usage{
board_colors(name = "default", overrides = NULL)
}
\arguments{
\item{name}{Preset name. Supported values are \code{"default"} and \code{"bs"}.}

\item{overrides}{Optional named list of color overrides.}
}
\value{
An object of class \code{backgammon_board_colors}.
}
\description{
Construct a board color preset
}
