% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{board_style}
\alias{board_style}
\title{Construct a board style preset}
\usage{
board_style(name = "default", overrides = NULL)
}
\arguments{
\item{name}{Preset name. Supported values are \code{"default"} and \code{"bs"}.}

\item{overrides}{Optional named list of style overrides.}
}
\value{
An object of class \code{backgammon_board_style}.
}
\description{
Construct a board style preset
}
